simplification of code for python 3 compatibility

+++ src/netius/common/calc.py
@@ -139,7 +139,7 @@ def egcd(first, second):
     if second == 0: return (first, 1, 0)
 
     q = abs(first % second)
-    r = long(first / second)
+    r = int(first / second)
     d, k, l = egcd(second, q)
 
     return (d, l, k - l * r)
